cpumanager: avoid calling updateContainerCPUSet() twice.
AddContainer() call already calls updateContainerCPUSet() for the
container. If the call to AddContainer() succeeded, we might as well add
the container to the success list and go to handle the next one.

The only thing we miss is that in case of an empty CPU set the container
is still put to the success list instead of failure list, since
AddContainer() doesn't report empty CPU set case as an error. The
comments say that the empty CPU set case should not happen in real world
though.