CPU manager support for QoS class based CPU sets.
If using a CPU manager policy different from None, take "be-cpuset"
and "other-cpuset" kubelet command line parameters in account when
assigning containers to CPU sets.

Modify AddContainer() and the reconciliation loop so that containers are
assigned only to the right CPUs (intersection of the assigned CPU set
and the QoS-determined CPU set. Modify Static policy so that the
available CPUs are coming from the CPU set provided by "other-cpuset"
command line parameter.