Add "be-cpuset" and "other-cpuset" parameters to kubelet.
The "be-cpuset" parameter controls the CPU set that will be assigned to
all containers whose parent pods belong to Best Effort QoS class.
Likewise, the "other-cpuset" parameters controls the CPU set for
Burstable and Guaranteed QoS classes.

The values for the parameters need to be CPU set "List" format (see 'man
7 cpuset'). For example string "1-3,5" would mean that the CPU set
would contain CPUs 1, 2, 3, and 5.

The default value for the paramters is the special value "all", which
means the CPU set containing all online CPUs present in the system.